need additional case to check for a lesser epoch
If isPkgShortNewer() was called with:

0:1.2.3
1:0.2.3

it would have returned true because the 1st package's epoch is
less but the version is greater.  Previously we only checked
for epoch >.  Also renamed the arguments to indicate that it is a
comparison of one package vs another and not a comparison of an
already thought of old vs new.